Move swap-domain styled components to appropriate folder
